changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11735 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/CommandCenter/ORCommandClient.h
@@ -21,6 +21,6 @@
 @interface ORCommandClient : ORSocketClient 
 {
 }
-- (void)netsocket:(NetSocket*)insocket dataAvailable:(unsigned)inAmount;
+- (void)netsocket:(NetSocket*)insocket dataAvailable:(NSUInteger)inAmount;
 - (void) sendCmd:(NSString*)aCmd withString:(NSString*)aString;
 @end
